pyproject.toml
[tool.poetry]
name = "attestation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
pillow = "^10.1.0"
djoser = "^2.2.2"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
